requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
bitarray==2.8.3
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
Flask==3.0.0
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
hexbytes==0.3.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lru-dict==1.2.0
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
parsimonious==0.9.0
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.1
py-solc-x==2.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
pyunormalize==15.1.0
pywin32==306
pyzbar==0.1.9
qrcode==7.4.2
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.1
rsa==4.9
setuptools==69.0.2
toolz==0.12.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
web3==6.11.3
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.3
